+++ src/simulation/continuous/brownian_bridge.rs
@@ -1,6 +1,6 @@
 //! Brownian bridge simulation
 //!
-//! For Brownian motion, see [`crate::simulation::bm`].
+//! For Brownian motion, see [`crate::simulation::continuous::bm`].
 
 use crate::{SimulationError, XResult, simulation::prelude::*};
 use rayon::prelude::*;
@@ -28,7 +28,7 @@ impl BrownianBridge {
     /// let mean = bb.mean(1.0, 1000, 0.1).unwrap();
     /// ```
     pub fn mean(&self, duration: impl Into<f64>, particles: usize, time_step: f64) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.raw_moment(1, particles, time_step)
     }
 
@@ -45,7 +45,7 @@ impl BrownianBridge {
     /// let msd = bb.msd(1.0, 1000, 0.1).unwrap();
     /// ```
     pub fn msd(&self, duration: impl Into<f64>, particles: usize, time_step: f64) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.central_moment(2, particles, time_step)
     }
 
@@ -75,7 +75,7 @@ impl BrownianBridge {
         particles: usize,
         time_step: f64,
     ) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.raw_moment(order, particles, time_step)
     }
 
@@ -105,7 +105,7 @@ impl BrownianBridge {
         particles: usize,
         time_step: f64,
     ) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.central_moment(order, particles, time_step)
     }