+++ src/simulation/continuous/levy_walk.rs
@@ -174,7 +174,7 @@ impl LevyWalk {
     /// let mean = levy_walk.mean(1.0, 1000).unwrap();
     /// ```
     pub fn mean(&self, duration: impl Into<f64>, particles: usize) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.raw_moment(1, particles, 0.1)
     }
 
@@ -196,7 +196,7 @@ impl LevyWalk {
     /// let msd = levy_walk.msd(1.0, 1000).unwrap();
     /// ```
     pub fn msd(&self, duration: impl Into<f64>, particles: usize) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.central_moment(2, particles, 0.1)
     }
 
@@ -224,7 +224,7 @@ impl LevyWalk {
         order: i32,
         particles: usize,
     ) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.raw_moment(order, particles, 0.1)
     }
 
@@ -252,7 +252,7 @@ impl LevyWalk {
         order: i32,
         particles: usize,
     ) -> XResult<f64> {
-        let traj = self.duration(duration).unwrap();
+        let traj = self.duration(duration)?;
         traj.central_moment(order, particles, 0.1)
     }