+++ src/simulation/continuous/mod.rs
@@ -1,15 +1,16 @@
 //! Continuous processes
 //!
-//! - Brownian motion in [Bm]
-//! - Fractional Brownian motion in [Fbm]
-//! - Generalized Langevin equation in [GeneralizedLangevin]
-//! - Langevin equation in [Langevin]
-//! - Levy walk in [LevyWalk]
-//! - Levy process in [Levy]
-//! - Ornstein-Uhlenbeck process in [OrnsteinUhlenbeck]
-//! - Subordinator in [Subordinator]
-//! - Inverse subordinator in [InvSubordinator]
-//! - Brownian bridge in [BrownianBridge]
+//! - Brownian motion [Bm]
+//! - Fractional Brownian motion [Fbm]
+//! - Generalized Langevin equation [GeneralizedLangevin]
+//! - Langevin equation [Langevin]
+//! - Levy walk [LevyWalk]
+//! - Levy process [Levy]
+//! - Ornstein-Uhlenbeck process [OrnsteinUhlenbeck]
+//! - Subordinator [Subordinator]
+//! - Inverse subordinator [InvSubordinator]
+//! - Brownian bridge [BrownianBridge]
+//! - Brownian excursion [BrownianExcursion]
 //!
 
 pub mod bm;
@@ -38,3 +39,6 @@ pub use subordinator::*;
 
 pub mod brownian_bridge;
 pub use brownian_bridge::*;
+
+pub mod brownian_excursion;
+pub use brownian_excursion::*;