+++ src/utils/mod.rs
@@ -4,12 +4,6 @@
 //! including mathematical functions, data structures, and computational methods specific
 //! to stochastic process simulations.
 //!
-//! # Components
-//!
-//! - [Auxiliary functions](functions) - Mathematical and computational helper functions
-//!   used in various simulation algorithms
-//! - [Circulant embedding](circulant_embedding) - Fast and efficient implementation of the circulant embedding
-//!   method for simulating stationary Gaussian processes with given covariance structure
 
 mod functions;
 pub use functions::*;